clean up tests and loading methods
