--- log.out.2017-08-16.0.log
@@ -1,2 +1,0 @@
-16.08.17 09:44:09 DEBUG [Jena                          ] 212   - <Jena initialization>
-16.08.17 11:15:25 DEBUG [Jena                          ] 212   - <Jena initialization>
