+++ pom.xml
@@ -3,20 +3,11 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.dbpedia</groupId>
 	<artifactId>KeywordSearch</artifactId>
-	<version>1.0.0</version>
-	<packaging>war</packaging>
+	<version>0.0.1</version>
+	<packaging>jar</packaging>
 	<name>KeywordSearch</name>
 	<build>
 		<plugins>
-			<!--Surefire - for JUnits -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-				<version>2.17</version>
-				<!--configuration>
-            			    <printSummary>true</printSummary>
-           			</configuration-->
-			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
@@ -25,20 +16,6 @@
 					<source>1.8</source>
 					<target>1.8</target>
 				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-war-plugin</artifactId>
-				<version>3.0.0</version>
-				<!--version>2.6</version-->
- 				<!--configuration>
-    				    <webXml>src\main\webapp\WEB-INF\web.xml</webXml>        
- 			        <configuration-->
-			</plugin>
-			<plugin>
-				<groupId>org.eclipse.jetty</groupId>
-				<artifactId>jetty-maven-plugin</artifactId>
-				<version>9.4.0.v20161208</version>
 			</plugin>
 		</plugins>
 	</build>
@@ -67,15 +44,12 @@
 			<version>4.12</version>
 		</dependency>
 
-		<!-- SPARQL Query Engine -
-			https://mvnrepository.com/artifact/org.apache.jena/jena-arq -->
 
 		<!-- For the F1-measurements -->
 		<dependency>
 			<groupId>org.aksw.qa</groupId>
 			<artifactId>commons</artifactId>
-			<version>0.4.11</version>
+			<version>0.4.13</version>
 		</dependency>
-
 	</dependencies>
 </project>
