+++ src/test/resources/file_sample_dbpedia_labels.nt
